test: add the first transmission-qt unit tests
