+++ tests/CMakeLists.txt
@@ -10,3 +10,7 @@ include_directories(
 
 add_subdirectory(gtest)
 add_subdirectory(libtransmission)
+
+if(ENABLE_QT)
+  add_subdirectory(qt)
+endif()
