+++ tests/helpers/Makefile.am
@@ -1,0 +1,4 @@
+EXTRA_DIST = \
+  sandbox.h \
+  session.h \
+  utils.h
