+++ tests/helpers/session.h
@@ -6,251 +6,30 @@
  *
  */
 
-#include "crypto-utils.h" // tr_base64_decode_str()
-#include "error.h"
-#include "file.h" // tr_sys_file_*()
-#include "quark.h"
-#include "platform.h" // TR_PATH_DELIMITER
-#include "trevent.h" // tr_amInEventThread()
-#include "torrent.h"
-#include "variant.h"
+#pragma once
 
-#include <chrono>
-#include <cstring> // strlen()
-#include <memory>
-#include <thread>
-#include <mutex> // std::once_flag()
-#include <string>
-#include <cstdlib> // getenv()
+#include "tests/helpers/sandbox.h"
+
+#include "libtransmission/crypto-utils.h" // tr_base64_decode_str()
+#include "libtransmission/file.h" // tr_sys_file_*()
+#include "libtransmission/platform.h" // TR_PATH_DELIMITER
+#include "libtransmission/trevent.h" // tr_amInEventThread()
+#include "libtransmission/torrent.h"
+#include "libtransmission/variant.h"
+
+#include <memory> // std::shared_ptr
+#include <mutex> // std::call_once
 
 #include "gtest/gtest.h"
 
-namespace libtransmission
+namespace transmission
 {
 
-namespace test
+namespace tests
 {
 
-auto const makeString = [](char*&& s)
-    {
-        auto const ret = std::string(s != nullptr ? s : "");
-        tr_free(s);
-        return ret;
-    };
-
-bool waitFor(std::function<bool()> const& test, int msec)
+namespace helpers
 {
-    auto const deadline = std::chrono::milliseconds { msec };
-    auto const begin = std::chrono::steady_clock::now();
-
-    for (;;)
-    {
-        if (test())
-        {
-            return true;
-        }
-
-        if ((std::chrono::steady_clock::now() - begin) >= deadline)
-        {
-            return false;
-        }
-
-        tr_wait_msec(10);
-    }
-}
-
-class Sandbox
-{
-public:
-    Sandbox() :
-        parent_dir_{get_default_parent_dir()},
-        sandbox_dir_{create_sandbox(parent_dir_, "transmission-test-XXXXXX")}
-    {
-    }
-
-    ~Sandbox()
-    {
-        rimraf(sandbox_dir_);
-    }
-
-    std::string const& path() const
-    {
-        return sandbox_dir_;
-    }
-
-protected:
-    static std::string get_default_parent_dir()
-    {
-        auto* path = getenv("TMPDIR");
-        if (path != NULL)
-        {
-            return path;
-        }
-
-        tr_error* error = nullptr;
-        path = tr_sys_dir_get_current(&error);
-        if (path != nullptr)
-        {
-            std::string const ret = path;
-            tr_free(path);
-            return ret;
-        }
-
-        std::cerr << "tr_sys_dir_get_current error: '" << error->message << "'" << std::endl;
-        tr_error_free(error);
-        return {};
-    }
-
-    static std::string create_sandbox(std::string const& parent_dir, std::string const& tmpl)
-    {
-        std::string path = makeString(tr_buildPath(parent_dir.data(), tmpl.data(), nullptr));
-        tr_sys_dir_create_temp(&path.front(), nullptr);
-        tr_sys_path_native_separators(&path.front());
-        return path;
-    }
-
-    static auto get_folder_files(std::string const& path)
-    {
-        std::vector<std::string> ret;
-
-        tr_sys_path_info info;
-        if (tr_sys_path_get_info(path.data(), 0, &info, nullptr) &&
-            (info.type == TR_SYS_PATH_IS_DIRECTORY))
-        {
-            auto const odir = tr_sys_dir_open(path.data(), nullptr);
-            if (odir != TR_BAD_SYS_DIR)
-            {
-                char const* name;
-                while ((name = tr_sys_dir_read_name(odir, nullptr)) != nullptr)
-                {
-                    if (strcmp(name, ".") != 0 && strcmp(name, "..") != 0)
-                    {
-                        ret.push_back(makeString(tr_buildPath(path.data(), name, nullptr)));
-                    }
-                }
-
-                tr_sys_dir_close(odir, nullptr);
-            }
-        }
-
-        return ret;
-    }
-
-    static void rimraf(std::string const& path, bool verbose = false)
-    {
-        for (auto const& child : get_folder_files(path))
-        {
-            rimraf(child, verbose);
-        }
-
-        if (verbose)
-        {
-            std::cerr << "cleanup: removing '" << path << "'" << std::endl;
-        }
-
-        tr_sys_path_remove(path.data(), nullptr);
-    }
-
-private:
-    std::string const parent_dir_;
-    std::string const sandbox_dir_;
-};
-
-class SandboxedTest : public ::testing::Test
-{
-protected:
-    std::string sandboxDir() const { return sandbox_.path(); }
-
-    auto currentTestName() const
-    {
-        auto const* i = ::testing::UnitTest::GetInstance()->current_test_info();
-        auto child = std::string(i->test_suite_name());
-        child += '_';
-        child += i->name();
-        return child;
-    }
-
-    void buildParentDir(std::string const& path) const
-    {
-        auto const tmperr = errno;
-
-        auto const dir = makeString(tr_sys_path_dirname(path.c_str(), nullptr));
-        tr_error* error = nullptr;
-        tr_sys_dir_create(dir.data(), TR_SYS_DIR_CREATE_PARENTS, 0700, &error);
-        EXPECT_EQ(nullptr, error) << "path[" << path << "] dir[" << dir << "] " << error->code << ", " << error->message;
-
-        errno = tmperr;
-    }
-
-    static void blockingFileWrite(tr_sys_file_t fd, void const* data, size_t data_len)
-    {
-        uint64_t n_left = data_len;
-        auto const* left = static_cast<uint8_t const*>(data);
-
-        while (n_left > 0)
-        {
-            uint64_t n = {};
-            tr_error* error = nullptr;
-            if (!tr_sys_file_write(fd, left, n_left, &n, &error))
-            {
-                fprintf(stderr, "Error writing file: '%s'\n", error->message);
-                tr_error_free(error);
-                break;
-            }
-
-            left += n;
-            n_left -= n;
-        }
-    }
-
-    void createTmpfileWithContents(std::string& tmpl, void const* payload, size_t n) const
-    {
-        auto const tmperr = errno;
-
-        buildParentDir(tmpl);
-
-        // NOLINTNEXTLINE(clang-analyzer-cplusplus.InnerPointer)
-        auto const fd = tr_sys_file_open_temp(&tmpl.front(), nullptr);
-        blockingFileWrite(fd, payload, n);
-        tr_sys_file_close(fd, nullptr);
-        sync();
-
-        errno = tmperr;
-    }
-
-    void createFileWithContents(std::string const& path, void const* payload, size_t n) const
-    {
-        auto const tmperr = errno;
-
-        buildParentDir(path);
-
-        auto const fd = tr_sys_file_open(path.c_str(),
-            TR_SYS_FILE_WRITE | TR_SYS_FILE_CREATE | TR_SYS_FILE_TRUNCATE,
-            0600, nullptr);
-        blockingFileWrite(fd, payload, n);
-        tr_sys_file_close(fd, nullptr);
-        sync();
-
-        errno = tmperr;
-    }
-
-    void createFileWithContents(std::string const& path, void const* payload) const
-    {
-        createFileWithContents(path, payload, strlen(static_cast<char const*>(payload)));
-    }
-
-    bool verbose = false;
-
-    void sync() const
-    {
-#ifndef _WIN32
-        ::sync();
-#endif
-    }
-
-private:
-    Sandbox sandbox_;
-};
 
 void ensureFormattersInited()
 {
@@ -479,6 +258,8 @@ protected:
     }
 };
 
-} // namespace test
+} // namespace helpers
 
-} // namespace libtransmission
+} // namespace tests
+
+} // namespace transmission
