+++ tests/libtransmission/bitfield-test.cc
@@ -6,10 +6,10 @@
  *
  */
 
-#include "transmission.h"
-#include "crypto-utils.h"
-#include "bitfield.h"
-#include "utils.h" /* tr_free */
+#include "libtransmission/transmission.h"
+#include "libtransmission/crypto-utils.h"
+#include "libtransmission/bitfield.h"
+#include "libtransmission/utils.h" /* tr_free */
 
 #include "gtest/gtest.h"
